cli/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "spectre-cli"
version = "0.0.9"
maintainers = [
  { name="Jimmy Fitzpatrick", email="jcfitzpatrick12@gmail.com" },
]
requires-python = ">=3.10.2"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "click<8.2.0",
    "typer-slim==0.13.1",
    "requests==2.31.0",
    "pyyaml==6.0.2"
]

[project.optional-dependencies]
test = ["pytest", "mypy", "black", "types-PyYAML", "types-requests"]


[project.scripts]
spectre = "spectre_cli.__main__:app"
